+++ app/controllers/admin/businesses_controller.rb
@@ -1,0 +1,7 @@
+module Admin
+  class BusinessesController < ApplicationController
+    def index
+      @context = BusinessesContext.new(title: "Businesses", search_query: params[:search_query])
+    end
+  end
+end