+++ app/controllers/admin/developers_controller.rb
@@ -1,0 +1,7 @@
+module Admin
+  class DevelopersController < ApplicationController
+    def index
+      @context = DevelopersContext.new(title: "Developers", search_query: params[:search_query])
+    end
+  end
+end