+++ app/models/stats/business.rb
@@ -1,0 +1,13 @@
+module Stats
+  class Business
+    private attr_reader :businesses
+
+    def initialize(businesses)
+      @businesses = businesses
+    end
+
+    def count
+      @count = businesses.count
+    end
+  end
+end